app/requirements.txt
tqdm==4.64.0
Pillow==9.2.0
scikit-image==0.19.3
